feat(spec): add beq instruction for the example RV128I emulator
Add detailed comments.
